common: removes duplicate compute_arch.* files
- it seems that the compute_arch.* files have not been removed following the refactoring
  effort commited as 84eb63f4 and have remained in use by various files throughout the
  code base. This commit removes the no longer relevant files from the common/internal
  folder and prepares the removal also from all include statements and other external
  dependencies

- Only sixtracklib/common/context/compute_arch.* should from here on be used!
